# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  ChatPresenceRequest:
    properties:
      media:
        example: text
        type: string
      phone:
        example: "5511999999999"
        type: string
      presence:
        enum:
        - composing
        - paused
        example: composing
        type: string
    required:
    - phone
    - presence
    type: object
  ChatPresenceResponse:
    properties:
      success:
        example: true
        type: boolean
    type: object
  CheckUserRequest:
    properties:
      phones:
        example:
        - "5511999999999"
        - "5511888888888"
        items:
          type: string
        minItems: 1
        type: array
    required:
    - phones
    type: object
  CheckUserResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/WhatsAppUserInfo'
        type: array
    type: object
  ContactDetail:
    properties:
      businessName:
        example: John's Store
        type: string
      jid:
        example: 5511999999999@s.whatsapp.net
        type: string
      name:
        example: John Doe
        type: string
      notify:
        example: John
        type: string
      verifiedName:
        example: John Doe Business
        type: string
    type: object
...
  ErrorResponse:
    properties:
      error:
        example: validation_error
        type: string
      message:
        example: name is required
        type: string
    type: object
...
  GetAvatarRequest:
    properties:
      phone:
        example: "5511999999999"
        type: string
      preview:
        example: false
        type: boolean
    required:
    - phone
    type: object
  GetAvatarResponse:
    properties:
      directUrl:
        example: https://...
        type: string
      id:
        example: "1234567890"
        type: string
      type:
        example: image
        type: string
      url:
        example: https://pps.whatsapp.net/v/...
        type: string
    type: object
  GetContactsResponse:
    properties:
      contacts:
        items:
          $ref: '#/definitions/ContactDetail'
        type: array
    type: object
...
  GetUserRequest:
    properties:
      phone:
        example: "5511999999999"
        type: string
    required:
    - phone
    type: object
  GetUserResponse:
    properties:
      jid:
        example: 5511999999999@s.whatsapp.net
        type: string
      pictureId:
        example: "1234567890"
        type: string
      status:
        example: Hey there! I am using WhatsApp.
        type: string
      verifiedName:
        example: John Doe Business
        type: string
    type: object
...
  SendPresenceRequest:
    properties:
      presence:
        enum:
        - available
        - unavailable
        - composing
        - recording
        - paused
        example: available
        type: string
    required:
    - presence
    type: object
  SendPresenceResponse:
    properties:
      success:
        example: true
        type: boolean
    type: object
...
  WhatsAppUserInfo:
    properties:
      isInWhatsApp:
        example: true
        type: boolean
      jid:
        example: 5511999999999@s.whatsapp.net
        type: string
      query:
        example: "5511999999999"
        type: string
      verifiedName:
        example: John Doe Business
        type: string
    type: object
info:
  contact:
    email: support@zpwoot.com
    name: zpwoot Support
    url: https://github.com/your-org/zpwoot
  description: |-
    A comprehensive WhatsApp Business API built with Go, following Clean Architecture principles.
    Provides endpoints for session management, messaging, contacts, groups, media handling, and integrations.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: zpwoot WhatsApp API
  version: 1.0.0
paths:
...
  /sessions/{sessionId}/contacts:
    get:
      consumes:
      - application/json
      description: List all contacts from the WhatsApp account
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetContactsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all contacts
      tags:
      - Contacts
  /sessions/{sessionId}/contacts/avatar:
    post:
      consumes:
      - application/json
      description: Get profile picture of a WhatsApp user
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Avatar request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/GetAvatarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetAvatarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user avatar
      tags:
      - Contacts
  /sessions/{sessionId}/contacts/check:
    post:
      consumes:
      - application/json
      description: Verify if phone numbers are registered on WhatsApp
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Phone numbers to check
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CheckUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CheckUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Check if users are on WhatsApp
      tags:
      - Contacts
  /sessions/{sessionId}/contacts/user:
    post:
      consumes:
      - application/json
      description: Get detailed information about a WhatsApp user
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Phone number
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/GetUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user information
      tags:
      - Contacts
...
  /sessions/{sessionId}/presence/chat:
    post:
      consumes:
      - application/json
      description: Send presence in specific chat (typing/recording)
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Chat presence request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ChatPresenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ChatPresenceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send chat presence
      tags:
      - Presence
  /sessions/{sessionId}/presence/send:
    post:
      consumes:
      - application/json
      description: Send general presence status (available/unavailable)
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Presence request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SendPresenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SendPresenceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send presence
      tags:
      - Presence
...
securityDefinitions:
  ApiKeyAuth:
    description: 'API key for authentication. Use the value from .env file (API_KEY
      variable). Send as ''Authorization: your-api-key'' (without Bearer prefix)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...